requirements.txt
PyQt5==5.15.10
pyqtgraph==0.13.3
pyserial==3.5
numpy<2.0
scipy==1.11.4
requests>=2.31.0
flask>=2.3.0
